ASM-8248 Fix rpm build for Makefile changes
A couple of recent ipxe upstream changes require include/assert.h and
include/ipxe/profile.h to be writeable in order to build. This change
updates those files to be owned by razor. It also updates the date to
match the date of the current upstream ipxe branch being used and the
ASM version to the current value (831).